Fix isNullish to not consider an empty string as Null.
An empty string is still a perfectly valid string, in both situations:
(1) An empty string is a valid string argument to a handler.
(2) An empty string is a valid non-null result from a handler.

I also removed the unused and unnecessary PrivateName value from the
NonNull struct -- it served no purpose and screwed up tests when isNullish
was fixed.